cheri: Use uintptr_t in tracing code
- Allow pointers in trace events (increases the alignment!)
- Use uintptr_t for pointers to trace buffers etc.
- Cast trace data as necessary.

+++ include/linux/fprobe.h
@@ -12,11 +12,11 @@
 
 struct fprobe;
 typedef int (*fprobe_entry_cb)(struct fprobe *fp, unsigned long entry_ip,
-			       unsigned long ret_ip, struct ftrace_regs *regs,
+			       uintptr_t ret_ip, struct ftrace_regs *regs,
 			       void *entry_data);
 
 typedef void (*fprobe_exit_cb)(struct fprobe *fp, unsigned long entry_ip,
-			       unsigned long ret_ip, struct ftrace_regs *regs,
+			       uintptr_t ret_ip, struct ftrace_regs *regs,
 			       void *entry_data);
 
 /**
